requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
jinja2
python-multipart
python-dotenv
